// src/app/dash/pages/services/interfaces/service-status.interface.ts
export enum ServiceStatus {
  pendiente = 'Pendiente',
  diagnostico = 'En diagnostico',
  espera_autorizacion = 'En espera de autorizacion',
  espera_partes = 'En espera de partes',
  reparacion_autorizada = 'Autorizado para reparacion',
  rechazado = 'Rechazado',
  reparado = 'Equipo reparado',
  listo_entrega = 'Listo para entregar',
  equipo_entregado = 'Equipo entregado',
  servicio_realizado = 'Servicio realizado',
  cotizacion = 'En cotizacion',
  cobro = 'En cobro',
  finalizado = 'Finalizado',
}